Scene Stuff
Scene overlay now properly works. Instead of loading the scene through SceneManager.LoadScene, it is added as an additive scene. The previous will be unloaded before changing over. The scene itself is set to Full HD (1920x1080) with the monitor surrounding it. All UI is set to world space. Cameras in the scenes are disabled except for the overlay scene aptly named for my internal screaming. A new script has been added for the MenuLoader object.