Reports Window Changes
Reports now properly scroll and change size depending on description length. This scrollbar also automatically updates the size as opposed to the prior implementation which required a second press of the report button. Reports without images now display the blank report background instead of lacking an image altogether.  An aesthetic address bar has been added to the window as this is a program affected by an internet outage.