Report Availability Overhaul
Reports now have IDs which correspond to the index of a bool list in LoadProgramManager. This list determines if the report can appear. They become unavailable when the report is answered 100% correctly. The list is reset once there are no longer enough reports to populate the day's list. Along with this, the remaining report images have been added, the report window has been enlarged, and a bandwidth monitor has been added to the internet panel.

All that remains currently is to resize the tutorial images (differently than in this commit probably) and to update the reports according to feedback.